CLDR-18402 CLDRConfig to always set CLDR_DIR=. with a warning
- .. if the current dir seems to be a CLDR dir
- we already do this for the .jar, but do this everywhere (such as from mvn:exec)
